rename: 파일구조변경 -#7
