+++ src/pages/join/index.tsx
